CRUD for specific user signed in, added some logging

+++ src/health-status/health-status.service.ts
@@ -21,8 +21,12 @@ export class HealthStatusService {
         return this.healthStatusRepository.getHealthStatus(filterDto, user);
     }
 
-    async getHealthStatusById(id: number): Promise<HealthStatus> {
-        const found = await this.healthStatusRepository.findOne(id);
+    async getHealthStatusById(
+        id: number,
+        user: User
+    ): Promise<HealthStatus> {
+        const found = await this.healthStatusRepository.findOne({ where: { id, userId: user.id } });
+
         if (!found) {
             throw new NotFoundException(`Health Status with ID "${id}" not found.`);
         }
@@ -36,16 +40,23 @@ export class HealthStatusService {
         return this.healthStatusRepository.createHealthStatus(createHealthStatusDto, user);
     }
 
-    async deleteHealthStatusById(id: number): Promise<number> {
-        const deleted = await this.healthStatusRepository.delete(id);
+    async deleteHealthStatusById(
+        id: number,
+        user: User
+    ): Promise<number> {
+        const deleted = await this.healthStatusRepository.delete({ id, userId: user.id });
         if (deleted.affected === 0) {
             throw new NotFoundException(`Health Status with ID "${id}" not found.`);
         }
         return id;
     }
 
-    async updateHealthStatusGenderById(id: number, gender: HealthStatusGender): Promise<HealthStatus> {
-        const healthStatus = await this.getHealthStatusById(id);
+    async updateHealthStatusGenderById(
+        id: number,
+        gender: HealthStatusGender,
+        user: User
+    ): Promise<HealthStatus> {
+        const healthStatus = await this.getHealthStatusById(id, user);
         healthStatus.gender = gender;
         await healthStatus.save();
         return healthStatus;
